requirements.txt
# Core dependencies
streamlit>=1.24.0
python-dotenv>=0.21.0

# PDF processing & conversion
PyMuPDF>=1.21.1
fpdf>=1.7.2
weasyprint>=59.0
markdown>=3.4.3

# Google API
google-generativeai>=0.3.0
